Rearrange context into app/context and separate LocalStorage provider
Circus doesn't seem to have a fail() equivalent: https://github.com/facebook/jest/issues/10953
