+++ app/experimentalFeatures.ts
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { FeatureDescriptions } from "@foxglove-studio/app/components/ExperimentalFeatures";
+import { FeatureDescriptions } from "@foxglove-studio/app/context/ExperimentalFeaturesContext";
 
 export default {
   diskBagCaching: {
